Corrects Bestiary Luck calculation based on tier
The Bestiary Actor's Luck is influenced in part by its difficulty,
and setting the Actor's "Only One Difficulty?" value to true
would not update that Luck value - it would maintain whatever
Luck value it received from its previous difficulty setting. This
updates the Luck calculation so that the difficulty value for the
Luck is always zero when the Bestiary Actor has only one
difficulty tier.