Adds new wounds calculations for 5.0 alpha
This updates the bestiary wounds calculation to use the formula
that Warhound passed on. I wasn't able to get it to the match the
values in the 5.0 alpha book, so I have to assume that this is the
formula being used to calculate bestiary wounds for the 5.0 beta.
When the beta rolls around, I'll revisit this to ensure compliance.

Additionally, this sees the Named Character wounds calculating
accurately with the new wording of Mythic Toughness, and with the
wording regarding how Jiralhanae and Hunters double their TOU mod
when figuring wounds.