Refactors weapon Item for special ammo format
This changes the weapon Item template to move all data that will
be relevant to special damage into a section `ammoList` under a
dedicated section `STD`. This will hold any of the weapon's base
stats that can be affected by special damage, and will serve as
an ammo type unto itself that every weapon will have by default.

This change necessitated refactoring all code that interacted with
the data as it existed before to point to the new data, including
the weapon sheet itself, the weapon summary, and all relevant
functions that calculated and reassigned derived values from both
`calculations.js` and `dice.js`.