Adds Cryo special rule to Weapon Items
This adds the new Cryo special rule to the Weapon Item. This
includes an update to the Weapon Item section in the template, a
new localization, acheckbox on the Weapon sheet to denote the
weapon as having the special rule, and a div to roll its damage in
the attack chat message.