Adds clickable link to special rule damage in chat
This makes each of the special rule damage labels in the attack
chat message clickable, which triggers a listener to calculate
the special rule damage based on the number of hits given by the
user in a dialog window (unless the special rule is Needle, which
just calculates a regular supercombine).

Next step is to `Roll` the damage and generate a chat message for
the damage roll using the `Roll` object's built in `render`
method.