Adds special rule links to post dmg rolls to chat
This adds links to each of the special rule damages posted in an
attack chat message. These use the formulas calculated in the
previous commit to create a `Roll` and pass the data along to the
template.

Included in this commit is a slight refactoring of the `dice.js`
module to move the `buildChatMessageData` and `postChatMessage`
functions to `chat.js` and export them from there.