Updates for beta release change to Rapid Reload
The beta has changed the Rapid Reload ability to no longer reduce
the reload Action required for some weapons to a Reaction; it is
now only able to reduce to 1 Half Action. This was updated here.

Not represented in the code is the fact that I also reviewed the
Wounds for everything in the Bestiary, and found that the formula
was indeed correct for the overwhelming majority. I had issues
with the calculations on Spartan IVs, Jiralhanae/Mgalekgolo,
Yanme'e, and Sharquoi. I will have to inquire with the developer
to fix these specific issues.