Adds extra check to not show medMgalekgolo skill
The medMgalekgolo skill was replaced with medXenophile some time
ago, and despite my best efforts the system seems intent on
checking for its existence whenever a sheet is updated. Presumably
the migration will stop this, but since I can't be certain, I've
added a check for it, and have the sheet skip over it if it finds
the skill on a character sheet. Worst case, this will mean a
character may have both medMgalekgolo and medXenophile, but only
medXenophile will ever be reachable.