Fixes negative characteristic bug for armor
The community pointed out a bug where armor could not handle any
characteristic penalties. The reason for this is because the armor
calculation logic would set any negative values to 0, and I had
not figured for values that could be negative like characteristic
and Mythic characteristic penalties. Now it checks a list of keys
that are allowed to contain negative values before deciding the
final value.