Adds Grunt and dev tasks
This adds Grunt as the task runner for the Mythic system and
introduces tasks to compile the Less.js files and copy all of the
essential system files to the Foundry systems directory, as well
as a file watcher.

One possible future modification of this would be to create a
release task that also unifies and uglifies all of the JS files,
minimizes the JSON and Handlebars files, copies everything to a
releases directory somewhere, and compresses it all into a zip
file to make ready for an automated release pipeline.