Updates Cauterize to 5.0 rules
This updates the Cauterize special rule to not have an input
field, and re-arranges a few fields in order to support a new
field `critsOn` that serves the same purpose that the
`cauterize.value` used to serve: determining what die value
constitutes a critical hit for the weapon.

The UI for the Weapon Item sheet changed as follows:

- The "Special Ammo Group" field was moved to the Special
Ammo tab.
- The "Price (cR)" field was moved to the space that was
previously occupied by the special ammo group.
- A new field "Crits On" was added to the basic weapon stats
table in the space previously occupied by the price.