Adds roll-to-chat functionality for known languages
This separates the original `postable` listener into two: one for
posting items to chat (`postable-item`), and one for posting
character details to chat (postable-details`). The new details
postable passes the entire actor to the template so it can decide
what data it needs.

This also moves the `language` class to the very top level so it
can be used in both the character sheet and the chat.