Adds rollable Education section to Actor sheet
This adds roll-to-chat functionality to the Educations section on
the Actor sheet. This required adding a new Handlebars helper that
can filter a list of items for the requested type, since the
item lists pre-filtered and stored on the `ActorData` is not
available outside the context of the Actor sheet.