Adds roll-to-chat function to Equipment
This adds the roll-to-chat for equipment items from the various
Actor sheets as well as from the Equipment Item sheet directly.

Worth noting here is that it does apply a link to the weapon and
armor items that are not working as of yet. These will be added in
future commits.