Adds roll-to-chat function to Weapons
This adds the roll-to-chat for weapon items from the various Actor
sheets as well as from the Weapon Item sheet directly.

Part of this required creating a function dispatched by the item
sheet to calculate the weapon's cost. It also required adding a
new Handlebars helper to check if one string is a substring of
another.

One field, "Mags Carried," currently doesn't exist. This will be
added in a follow-up commit, which will also fix this field to
read the actual value.