Adds magazines carried value to Weapon Item
This adds a new number field `magsCarried` that will keep track of
how many loaded magazines the character is carrying for the weapon.
Reloading a weapon will automatically decrement this value by 1,
warn the user when they are on their last mag, and interrupt a
reload action when they have insufficient magazines to complete the
reload.

When special ammo is implemented, this value will be held on each
special ammo object in a weapon's `ammoList`.

It's worth pointing out as well that this commit includes a
correction to the `makeUIWarning` function to call the correct
Foundry function to propagate a UI warning.