Restoration and minor refactor of magazine count
This restores the magazine count logic, and refactors it so that
all of the actual business logic occurs in one location as opposed
to being performed in the callbacks on each Actor sheet.

The next steps will be to add a new system setting for the user to
choose between a magazine counter or an ammo pool, to modify the
`calculateReloadMagFed` function to account for this setting, and
to write `calculateReloadSingleLoading` using the ammo pool logic,
as well as to account for a "magazine 0" weapon like the homebrew
bow.