Adds ammo pool reload logic
This adds the logic to manage reloads for weapons that use an ammo
pool system (single-loading weapons as well as ALL weapons when the
magazine mode setting is active).

This also adds a plus (+) in front of the single-loading reload
value on the weapon summary partial to visually differentiate it
from the number of half actions that magazine-based weapons use.