Refactors ammoList reference to use current ammo
This adds a new field `currentAmmo` to the weapon item that will
eventually store the currently selected special ammo loaded into
the weapon. For now, it defaults to "STD" so that it can reference
the base weapon stats.

All existing references of the old logic in the JS modules have
been replaced with the use of this field.