Adds value field to Electrified Special Rule for Weapons
Unfortunately, this had to be forced to display in the Weapon
sheet with a conditional checking if the special rule in the table
had this name. The migration wasn't working, and I can only assume
this is because of the move to Foundry v10. I've added a TODO to
remind myself (hopefully) to fix this after the move to Foundry v10
has been merged in.