Changes AGI mod for charge/run field and calculation
Previously, Hunters gained double their AGI modifier when figuring
Charge and Run speeds. Instead, they now gain a +3 bonus to their
AGI mod for the purpose of figuring these movement speeds.

This required changing the `doubleAgiRunCharge` value to the new
`agiBonusRunCharge`, which takes an integer instead of a boolean.
This is a more dynamic field that allows for both the old method
of calculating this ability for Hunters, as well as the new way
and possible future official or homebrew abilities that give a
similar benefit with different values.