Splits calculation of stats into two groups
The first batch is only STR and TOU. The second batch is all of the
remaining characteristics. This allows us to calculate the two
characteristics that are important for figuring carrying capacity,
then calculate inventory weight and carrying capacity to determine
if the character is overencumbered so that the appropriate
penalties can be applied to AGI.

This also moves the calculation of inventory weight between the
calculation of these two groups of stats, and moves the calculation
of the inventory bar values to after the calculation of carry
weight. This puts the new process in place to proceed with the
calculation of penalties.