Perform stat calculations in more functional way
The previous work done to "split" calculation of characteristics
was misguided. No one characteristic is dependent upon another for
its calculation, and carry weight is dependent on the calculation
of Mythic characteristics which were calculated afterward anyway.

Instead, this is changing the `calculateCharacteristics`,
`calculateMythicCharacteristics`, and
`calculateCharacteristicsFlood` functions to perform their work in
more a more functional way, and does some reorganization of when
these functions are called in order to make ready for the next
commit which implements `calculateEncumbrance`.