Preps for adding encumbrance calculation
The prior simplification of `calculateCharacteristics` was a bit
premature - in order to use a penalty on the `roll` value for AGI,
we will have to calculate those values AFTER assigning the penalty
which means splitting the characteristic calculations and doing the
encumbrance calculation in the space between.

Part of this was also adding a setting to switch between using
Mythic's typical rules for encumbrance, turning encumbrance off,
and adding my personal simplified rules for encumbrance.