Empties bogus armor shield and stats on Actor
Up to now, armor was applying its modifiers for shields and
characteristics to the worn Actor even if the `has` field was not
checked. This was caused by the function to apply these to the
Actor only updating these values when that field was checked, but
we are still doing the math on the existing fields elsewhere. This
commit now empties those fields on the Actor so they won't be
applied at that stage of the calculations.