Adds per-tier luck input field
In 5.1, there is no consistent pattern for luck advancements in the
Bestiary. To allow for accurate luck tracking across difficulties
for Bestiary enemies, this adds a number field for each difficulty.
The box corresponding to the currently selected difficulty for a
Bestiary Character Actor is styled bold for easy recognition.