Adds new experience payout fields and logic
This updates the experience payout to be 5.1 compliant. Rather than
multiplying a base value by the difficulty tier, there are now
individual fields for each difficulty to put the base value in. We
must do this since 5.1 has de-standardized the calculations of
experience payouts. The additional experience payout value for the
kit is kept in. Flood characters are unaffected by this change.