Moves scatter dialog options to attack dialog
This requires the distance to target and zero-G combat flag in the
attack dialog rather than in the scatter dialog. This prepares for
using these values to calculate distance penalties and bonuses for
attacks, as well as for calculating scatter at the time when a
qualifying attack misses.

While it is no longer being used, the scatter dialog template will
stick around since it will eventually be used in a scatter macro.