Performs scatter calculations at time of attack
This changes the code to determine an attack's scatter, if any, at
the time when the attack is determined to have failed. This allows
us to store the scatter data and persist it between server
instances. Now, when a scatter is performed by the user, it simply
displays the existing scatter value.