Updates equipment trainings to new list
This updates the equipment training groups and descriptions to the
new 6.0 values. Unfortunately, this results in a breaking change
due to the way that the code loops over the equipment groups in
the Actor template. All attempts at migrating the sheet have
failed. All other breaking changes should be made at the same time,
including the characteristics and the v10 migration.