Adds range modifiers to weapon attacks
This uses the "Distance from Target" field added to the attack
dialog to calculate attack, damage, and pierce modifiers whenever
an attack is made, and applies the values to the result of the
attack as appropriate. This feature lives behind a system setting
that allows users to disable it at any time.

This still does not apply penalties when a target is outside the
attacker's perceptive range - this will require adding a scope
magnification field to the Weapon, which will be done in a future
story.