Moves special rules into STD ammo object
This removes special rules for a weapon by default, and instead
places all the weapon special rules onto its `STD` ammo object.
This will allow for the weapon's effective special rules to change
with the selected ammo, as well as making it easier to figure its
special rules once special ammo selections are in play.