Rolls damage at time attack is made
Rather than doing the damage roll in-line by passing the damage
roll formula in square brackets to the chat message directly, this
rolls and interprets the damage roll at the time the attack is
determined to be successful (or the result of an AOE attack). The
style of the roll result rendering hasn't changed, but since we do
have access to the roll result, we can perform work on that result
to determine special damage. This will be implemented sometime in
the future, but for now we are rendering a bullseye icon when the
attack would apply special damage - that is, rolls any die result
meeting or exceeding the weapon's `critsOn` value.