Fetches data from Google Sheets and converts it
The `main` function now successfully fetches data from Google
Sheets and converts each line into a `RawEquipment`. Note that
this is currently only working for the first section of the
equipment sheet - future commits will work on splitting the CSV
data into separate compendia-based CSVs (with the header) and
dispatching those for conversion.