Implements main loop for program
The program will primarily work by traversing over a list of sheet
subjects, running the main program for each in turn and collecting
the results as `Either`s. This will allow it to short circuit in
advance if something goes wrong with one of the sheets. We have
defined the ranges and GIDs for each sheet as a `Map` for easy
lookup and traversal.