Switches to compendium content map
This switches from using a newtype declaration to pair compendium
metadata with its content, and instead uses a `Map` of that data
to its content instead. Now, each sheet will carry a map of its
various compendia to their entries, which has already simplified
the code returned and made it easier to relate the data.