Converts Foundry data to compendia
This converts each Foundry item from the previous step in the
process into an `Entry`, and packages them all as part of a
`Compendium` - one for each sheet. Testing shows that the sheet
splitting isn't working, and will need to be revisited.

What has proven to be working, though, is the item ID generator.
This uses the item name - that is, the combination of the label
for the compendium and the name of the item itself - and uses that
as the seed for a pseudo-random number generator which picks 16
alphanum characters. Because the seed is consistent each time, as
well as unique to each item (this can be reasonably guaranteed),
we can reliably generate the same ID for each item every time the
packs are generated. This would allow for referencing other items
in the descriptions of future ingested items outside the Domain.