Fixes sheet compendia separation
This fixes the separation of the equipment sheet pulled down from
Google Sheets into compendia. It requires a really gross parser,
but I don't have any better answer for the problem at hand at the
moment.

Consequently, this did allow for the removal of attoparsec, but I
suspect it will be back when it comes time to do the conversion
code for the other Foundry items.