Makes main operation asynchronous
This handles the fetching and transforming concurrently, since
each is only inside of `IO` in the first place because of the HTTP
request to fetch the Google Sheet.

We're also exiting the entire program whenever the result of any
conversion operation returns a `Left` by calling `System.Exit.die`.

Lastly, we're ignoring the argument in the `ToJSON` instance for
`Breakpoints` since they aren't implemented and are defaulting to
all-zero values anyway.