Collects compendia for later manifest encoding
Rather than discarding the results of writing the compendia .db
files to disk and exiting, we will now instead hold onto the
compendia so they can be pretty-encoded to JSON for the manifest
files.

Another consideration: should we encode the entire system manifest
so we don't have to write the raw JSON somewhere and copy it into
the manifest? Instead, we can write the prettified JSON directly
to a manifest file.