Encodes compendium entries into compendium files
We are now successfully printing (minified) entries into a .db file
and storing them in the appropriate directory. They are printed as
one entry per line, rather than as proper JSON, as Foundry requires.

The next step is to encode the array of pack objects for the system
manifest. Once that is done, we can test.