Rewrites entire manifest file when run
This re-writes the entire manifest file to disk each time as part
of the main loop instead of just writing the pack JSON to an
otherwise useless file. Part of this was adding manifest-related
data types which will be useful later if The Domain gets widened to
be a community tool.