Removes Error throwing from dice roll inerpreter
An issue came up for a user that may have been related to the dice
roll interpreter function. My theory is that a bad character made
its way into the field, causing an unhandled error to be thrown
that broke some piece of the code/exited the Actor calculation
process too soon. I can't verify this, but this function shouldn't
be throwing unhandled exceptions anyway. Now, it simply returns an
empty set of modifiers.