Extracts shields to templates in data model and HB
The shields object has been extracted to a template that can be
applied in the `template.json` module, and was applied to armor,
equipment, and weapons. The Handlebars code for the shield table
has also been extracted to a partial, and applied to the armor
sheet to bring everything back to how it was. The next update will
add that partial to each of the other applicable item sheets.