Adds new fields for melee and thrown weapons
This adds a new field Extra Full Attacks for melee weapons and a
new field Range Bonus for thrown weapons. These fields help to
balance out the table against the ranged weapon's render of the
table which requires more cells, as well as granting flexibility
with these items to add flat bonuses to the indicated elements of
the weapon's operation.