Adds shields partial to equipment sheet
This adds the shields table to the equipment sheet so that shield
values for equipment items can be declared. This required a small
refactor to how some of these values are calculated for items,
which is going to make it more flexible as any type of item with
these values can map over them and calculate them in the same way.