Adds shields summary table under weapon summary table
This table holds all of the equipped non-armor items the character
possesses, and displays their relevant shield values and allows
the character to recharge each shield-granting item individually
as well as to recharge all of them at once. This does not currently
show shields granted by armor - it is meant to track extra shields
granted by equipment like a bubble shield or a weapon like the
point defense gauntlet. Armor is still tracked as before.