Fixes bad state around setting encumbrance penalty for fatigue
As it turns out, the issue was related to fatigue, as I had assumed
the first time this came up. The penalty was always exactly 5, which
is the multiplier for characteristics penalties due to levels of
fatigue. However, it wasn't related to any visible level of fatigue,
but instead coming from the point of fatigue caused by encumbrance.

When recalculating an Actor for the first time, it can get into a
bad state if the first thing done on the Actor sheet (and thus the
trigger for the first-ever model re-building) isn't to assign STR
or TOU a positive value. In this case, it notes that the character
is fatigued because they're carrying an amount of weight (empty
inventory; 0 kg) equal to their carrying capacity (0 STR + 0 TOU;
0 kg), but nowhere can it un-encumber the Actor by flipping that
value back.

The solution applied here was to set the encumbrance to `false`
explicitly during the calculation of inventory weight. This is a
shared function that all three Actor types call during the
`prepare_Base` stage. It's tangentially related to encumbrance, so
it makes enough sense to put it there. A better solution would be
to do more explicit logic around the assignment of that encumbrance
penalty toggle for fatigue, but switching it for all cases would be
cumbersome, and I'm not even convinced that it would be a better
and more clear way of doing it. And, given that I want to move to
the 100DOS implementation ASAP, I'd rather not waste time on it.