Parses shields out of equipment descriptions
While brittle, parsing the free text descriptions for each
equipment item is the only way to determine whether it has or
grants shields. This is an attempt to parse those values out for
compendium pack items, and is currently parsing these values out
correctly for all 4 known equipment items that have or grant a
shield. This will not be able to recognize when the text changes
for these items and thus the parsing breaks, so any future loss
of this data for the equipment items will have to be called out by
a user or forgotten.