Adds debug module
This introduces the `Domain.Debug` module, which currently only
exports a function `writeDebugFile` which writes the .csv file
as received from the request to Google for review when debugging
parse failures.

It is meant to be called from `Main` before the main conversion
sequence, and will only be imported to make the debug file. That
is why the module exists, but is not imported anywhere in this
commit.