Uses a sheet preparation context type during sheet prep
When splitting a sheet into its own compendia, we were originally
passing around a lot of confusing arguments to a recursive function
that made understanding what was happening difficult. This instead
adds a new `Context` type to the `Domain.Prepare` module and passes
and modifies that with each line parsed. This is much easier to
follow, and would make it trivial to add additional arguments
during this process in the future if required.

Part of this includes adding some newtypes and type aliases to
clarify the intent behind similar types within that `Context`,
and a helper function to handle pushing a new compendium to the
accumulated `CompendiumMap`.

It also adds constants for all of the different sheet subjects'
empty compendium details. They still require values to be added,
which will be done when those item types' compendium builders are
implemented. They're being added now just to prevent compiler
warnings around an inexhaustive case expression in `prepareSheet`.