Simplifies sheet splitting logic
This ignores the compendium title, instead using the sheet subject
to determine the compendium title and only splitting to a new
compendium when the faction changes. We can safely determine, for
now, that the faction is always listed in column 3 in hidden lines.

This results in simpler, easier to understand code, at the cost
of a smaller number of larger compendia. I have updated the
planned compendia in the `README` as a result.