Parses special rules out of weapon tags
This splits the content in weapon tags into possible special rules
and possible tags. This is done by splitting the tag contents into
words, taking the last word as the possible tag, and re-combining
the remaining words as the potential special rule. If this isn't a
proper special rule, it will be added to the description as per the
previous commit.