Refactors skills to iterate over config skills
Rather than iterating over the skills in the Actor's skills object,
we're now iterating over the skills listed in the config. This will
make it easier to handle the renaming or removal of skills in the
future, since it will no longer be printing old skills that still
live on an Actor's data model.