Fixes characteristic test flavor text
The change to Courage, Charisma, and Leadership's abbreviations
broke the flavor text lookup for tests on these characteristics
because the abbreviations were being used as the key values in the
constant `CHARACTERISTICS` dictionary and they weren't updated at
the time the change was made. I would like to remove this at some
point and just rely on `localize` but that proved too complicated
for now due to the way it's being used to look up the message's
title in `getTestOptions`.