Adds button to create items from Actor sheets
This adds a button that creates an Item directly on the various
Actor sheets rather than requiring the user to drag items onto the
sheet. This works for all Item types, creating them with their
minimal complete definitions so that they display correctly on the
sheet upon creation. It has been tested and shown to work on all
three Actor sheets for all Item types and subtypes.