Simplifies Foundry item parsing
With item factions now a column in the spreadsheet, they can be
parsed from the CSV along with the other data. This eliminates the
need to try and split the CSV rows by faction in order to organize
the compendia. Now, we simply filter out "empty" CSV rows prior to
ingestion.