Removes unnecessary field from RawEquipment
Previously, we were dependent on the `COMP_field` column of the
equipment spreadsheet to make decisions about how we split the
sheet into multiple documents for parsing into separate compendia.
Now that the faction is being added as a column, we're splitting
the compendia by the faction on the parsed weapon, making this
field no longer required.

This also updates the range of the equipment spreadsheet in the
request parameters to satisfy the addition of a new column and to
fetch the weapon mods which we previously couldn't parse.