Stubs out data model for Vehicle Actor
Based on an initial first pass at the Vehicles section, this lays
out the apparent elements of vehicles in Mythic 6.0 as they will
have to be represented. This structure may change as development
continues.

Within this data model lies three implicit features we intend to
support:

- Automated adjustment of movement and characteristic values based
on loss of propulsion.

- Communicating the vehicle's "doom state" to the user based on its
current damage.

- Roles that can accept Actor ID values to populate and modify the
vehicle's maneuver, evasion, stunting, and weapon attack values.