Adds internal labels to vehicle details input fields
This adds labels to the interior of the input fields for vehicle
details number-type inputs. This is a huge UX improvement over the
unlabeled fields that would rely solely on placeholders since it
actually communicates what the field holds rather than just a lone
number value. Once the vehicle work gets merged in, this will be
expanded to cover like fields in the headers of other sheets, too.