Wireframes vehicle gauge and stat block
This adds wireframing for the Vehicle sheet summary tab, except for
the weapon summary section. This includes the speedometer/gauge
from the TSL module, as well as (non-functioning) doom and splatter
buttons, energy shields with recharge button, and conversions for
current and top speed. Another section will be needed for walker
vehicle stats; it will replace the speed gauge for applicable
vehicles.