Adds weapon summary section to Vehicle sheet
This adds the weapon summary section to the Vehicle sheet as the
final part of its summary page. Rather that creating a new, nearly
identical partial, it made more sense to add the conditions for a
Vehicle-specific render since there are only three:

- Full colspan for the table header
- Don't render the initiatve and evasion buttons
- Render the crew role select element for each weapon

Incidentally, this also resulted in having to style the weapon
name above a pair of selects, the second (currently nonexistent)
one being for the current ammo selection. When this is added, we
will have to add a fourth condition to render that with a value of
`span 2` for `grid-columns` since it will be offset without the
role select.