Adds Vehicle sheet Settings basic info
This introduces a Settings tab for the Vehicle sheet to manage a
vehicle's base/max values for breakpoints, armor, shields, size
points, weapon points, propulsion, and movement or characteristics
depending on propulsion type. Crew/complement role assignment is
not included in this, as it will be done in its own story with the
event listeners later.