Adds Cargo tab to Vehicle sheet
This adds the fourth and final tab "Cargo" to the Vehicle sheet to
manage carried and equipped items and weapons. It's mostly a
simplified version of the Inventory page from other Actor types.

It’s worth mentioning that, in order to get the weapons in the
summary section to render without errors, perceptiveRange had to be
added to the Vehicle Actor despite their not having such a value.
It also required adding owner to the weapon item - this is
populated from the dropdown, and will be used at the start of the
weapon preparation process to be passed as the actor wielding the
weapon.