Implement vehicle propulsion calculation
We can now determine the reduction to a vehicle's movement based on
the ratio of active to total propulsion units. This required a new
way for the user to input max propulsion units - a select menu of
fixed numbers of units defined in the config for everything but the
thrusters, which remains the same. This was all implemented in a
vehicle-specific module, to which we've also extracted the doom
calculations.