Fix inverted doom state
Vehicle doom state was being inverted due to the negating of the
hull breakpoints occurring twice: once in the `calculations` module
and once in the `vehicle` module. Realistically, we want all of the
logic concerning vehicle calculation to occur in the `vehicle`
module, so it should be the only one that performs the negation of
hull breakpoints.