Updates vehicles when one of their crew is updated
During the data prep stage for any non-vehicle Actor, we now check
for any Vehicle Actors with that Actor assigned to a crew role,
and force an update and re-render on any that we find. This makes
it so changes to a crew Actor will visually update  the Vehicle
right away, even if the sheet is opened and without requiring a
change on the Vehicle sheet.