Vehicle weapon attack stats derived from Actor of assigned role
The role select for the weapons on the weapon summary partial are
now updating the weapon's `owner` value. This allows us to use it
when calculating the weapon's attack stats, using either it or the
parent Actor as appropriate.

This did reveal a bug where select options that have a `value`
attribute of some string that _begins_ with a number will be set
to that number, at least in Foundry. This required an alternate
approach to getting the Actor ID into that attribute, which will
have to be applied to Maneuver, Evasion, and Parry values for the
Vehicle.