Adds wreck event listener
This adds an event listener for vehicle wrecks and collisions. For
the most part, it's the same as the splatter post without the
penalty to Evasion tests. Both of these listeners are minimal for
now, but there is room to expand them later since no additional
information is required for the Vehicle data model.