Sets weapon target to 0 when no owner, disables attack buttons
When calculating weapon attack summary data, we now set the target
value to 0 when no owner can be determined. This usually indicates
that the weapon is not assigned to a role, so we don't want to
permit it to be fired. So, we also disable the attack buttons when
the target number is 0.