Fixes rich text editor button spacing bug
This was caused by a width of 85% being set for all buttons. Since
the rich text editor tool options are button elements, this caused
them to take up 85% of their entire row, which made them not fit
on the same row in a horizontal stack, and forced new rows to be
made for each button.