Fetches Actor perceptive range data within Handlebars templates
This introduces a Handlebars helper that can fetch Actor data from
a provided ID. This is being used to fetch Actor data for the owner
of a Vehicle Weapon, which would otherwise be unavailable in the
context of the weapon summary partial. It's being used to calculate
the maximum perceptive range for ranged weapons.