Fixes location switching for attacks
Previously, we had a field on the attack data that's passed along
through the attack process that would indicate whether the target
was a vehicle (and therefore required vehicle hit locations on a
successful attack) or a normal character. Somehow this value got
renamed to a bad reference, so it was always set to `undefined`
and treated as `false` by JavaScript. This puts the correct name
for the reference back.