Adds Ability compendium
This creates a compendium for Ability Items. This is (likely) the
only factionless compendium that will be integrated, so one aspect
of this was adding nullability for the `Faction` portion of the
`CompendiumData` context that's carried with the parsed sheet data.

This has been tested and shown to work, with two noteworthy points:
  - It was only tested up to a certain line due to multiline parse
issues.
  - The items currently do not have an image set.