Moves FoundryData builders to use Either monad
This takes advantage of the `Monad` instance for `Either` and the
`maybeToEither` function to avoid having overloaded copies of the
builder functions essentially acting as guards for the `Faction`
maybeness. I find this easier to read and it's already established
in the event that more terminal failures to make entries come up.