Ingests entire Abilities sheet into a usable compendium
This ingests a duplicate version of the R20 Abilities sheet in the
Mythic Dev Sheet. This version has the experience and description
columns sanitized to remove illegal characters, such as commas in
the experience column and newline characters in the description
column. This fixes the multiline parse issues and allows the entire
sheet to be ingested.