Sets attack count to 0 when magazine is 0 or less
This sets the attack counts to the greater of 0 or their value
based on magazine count, instead of just the mag count when it was
less than the maximum potential attack count. This prevents the
negative attack count that wasn't accounted for when the magazine
is a negative value.

It also disables the single attack button when the magazine is not
greater than 0, instead of when it is just 0.

Admittedly, this situation should never come up in actual practice,
but players could potentially set the magazine count to negative,
so it should be guarded against.