Extracts ranged weapon variants into their own raw ingestion type
Due to the latest change to the ranged weapon compendium spreadsheet
putting up to three variants on the same line under the same weapon,
a rework was required that would produce a list of weapons when
converting the raw data instead of just a single weapon. This, in
turn, required the raw data parsing out a list of variant data all
originating on the same row.

We accomplish that here by extracting all the data appearing on the
variants to a new type `RawRangedBase`, and parsing a `NonEmpty` of
that type from the base (0-index) data and an optional second
variant if there is data in that variant's name field. The function
that converts from raw data to the Foundry data now returns a list
of the Foundry types and concats them at the end of the traversal.

Testing for this has gone far enough to confirm that the shete is
being ingested into the raw data type successfully, but the function
converting raw to Foundry data is just returning an empty list now.
The next step will be to perform the actual conversion by modifying
the commented-out conversion code to work with the new structure.

If this is successful, the same approach will be extended to melee
weapons.