Adds Flood bestiary compendium
This introduces a compendium of Flood Actors. This is the first of
the two planned Actor compendia. Some refactoring was required to
support adding items to an entry, as well as to differentiate
between what type of Ability was being created. A considerable
number of new types were added to support Actors as well, which
weren't previously necessary since they aren't shared with any
Item type.